+++ History.markdown
@@ -1,5 +1,7 @@
 ## HEAD
 
+  * define path with __dir__ (#109)
+
 ## 2.8.0 / 2017-08-15
 
 ### Minor Enhancements
@@ -19,7 +21,7 @@
   * Set title and description in dev (#104)
   * Detect whether the client is connected to the internet. Only allow client calls if connected. (#90)
   * Expose repo visibility (#105)
- 
+
 ## 2.5.0 / 2017-07-17
 
   * Set default `site.title` and `site.description` (#101)